Complete UI composition refactor round
